Mega-trends Driving Chemical Industry -Pt.1
                  Mega-trends Driving Chemical Industry - Digital Strategies. Pt.1
Introduction.In these challenging times, the chemical industry has experienced its vulnerability and shown its tenacity. Disarray is a continuous thing and is the reason for the evolution of an industry in accordance with market changes and customer needs. To gear up to these challenges and to adapt to digital technical advancements, new levels of agility, scalability, data leveraging, collaboration should be adapted. Opportunities to lead the future of innovation will only present to those corporations that can integrate, simplify, and automate processes while bringing in more sustainable business models. Humans are the key enablers to all these transformations, which are technology intensive. End-to-end process reexamination will yield significant cultural changes in chemical companies. These processes extend beyond the enterprises. The industry will emphasize a common platform to share information with those across the value chain, earning trust and laying the foundation for a more sustainable future. The path of continuous innovation and leverage existing technologies will lead the Chemical companies to be the leaders in Digital transformation journeys. For instance, the use of artificial intelligence, automation, and optimization can unearth new sources of value while empowering business users to focus their time and energy on high-value activities.
The adoption of industry 4.0 is the key element for any industry which is underway in its digital transformation as intelligent technologies are embraced, and new levels of connectivity are brought into historical and siloed data. New business models based on real-time data sharing, and co-innovation with customers on new platforms and in ecosystems, will continue to be strong differentiators in the chemical industry. These innovations or opportunities for innovation include, for example, sharing data with customers to develop custom products, using real-time data to diagnose and correct production issues as they happen, and developing business models based on value rather than volume. Industry 4.0 is only one area where new digital technologies make it possible to gather and analyze data across machines and business systems to enable faster, more flexible, and efficient processes.
All these changes and innovations are driven by megatrends which can be categorized into three main sections:
•	Digital Strategies of Transformation
•	Success Priorities of Transformation
•	Underlying Technologies of Transformation
We will cover all three megatrends in this series of articles. The current article will majorly focus on ‘Digital Strategies of Transformation’.
Six Digital Strategies of Transformation
The Six Digital Strategies of Transformation which will decide the leaders for tomorrow are as follows:
1.	Increasing competition is driving the commoditization of products and the erosion of margins. This is resulting in chemical companies focusing on co-innovation, portfolio optimization, and selling of outcomes and business values instead of just developing products with the ultimate goal of generating new customers and consumer experiences.
2.	Sustainability is the new digital transformation that affects all areas of business. From reducing the overall cost of the products by decreasing the cost of operations through efficient energy management, to getting subsidiaries from different governments to drive the sustainable model, all while increasing the profit margins.
3.	The risk of supply chain breakage has increased dramatically in the past few years. The breakage ranges from pandemics affecting the availability of human resources to support operations, to wars disrupting international trade, taxation and various other increased risks. Various geopolitical risks, waves of pandemics and climate change play a crucial role in risk management now.
4.	Intelligent Technologies and automation are the new definitions of technical advancements. Technologies like the Internet of Things, Artificial Intelligence, Intelligent robotic process automation, machine learning, blockchain, the cloud and analytics, provide new opportunities for chemical manufacturers to cut costs by automating the back office and running low-touch operations.
5.	Starting to Participate in Business Networks is a step beyond traditional value chains for chemical companies, as they now play in the field with disruptors coming from all angles. Hence, thriving along with business partners becomes the right choice for them.
6.	In an increasingly dynamic world with mergers, acquisitions and divestitures, as key vehicles for portfolio optimization and sustainable growth, strategic Market driven agility has become imperative to survive and thrive.
Digital Strategies are ever evolving and transforming, and by this they also change the rules to be the leader in the market for Chemical Companies.
<aside> 💡 Studies suggest that about 45% of the chemical companies in the market are rethinking about the ways their human and machine resources are interacting with each other. 94% of executives in the chemical industries survey ways to increase their company’s digital expenditure and bring in the new positive transformation. 79% of industry experts believe that the best ROI will come for a company having either Cloud Computing (45%) or Big Data Analysis (34%).
</aside>
The ability to address these global trends will determine who will be among the winners in the midterm and long-term future. According to Accenture, tomorrow’s leading companies are already moving beyond products and services. They are applying technology to create deeper and more meaningful relationships with people. They are creating new affiliations with businesses across industries that share their vision and mission. They are using these new partnerships to invent new products and services that meet the goals of their customers and employees and, in doing so, are achieving new levels of growth and differentiation. They are also helping their communities create new economic opportunities and develop new ways to serve and protect citizens, benefitting society as a whole. Industry 4.0 is an enabler to achieve this, leading to a true Fourth Industrial Revolution.